requirements.txt
pymupdf
requests 
numpy 
faiss-cpu
streamlit 
